+++ tests/asm.rs
@@ -303,7 +303,14 @@ fn test_str() {
         Pop(None, 1),
         Dec(D),
         End,
-        CoreOp::stack_alloc_cells(C, vec![0].repeat(1024)),
+        Global {
+            name: String::from("testing"),
+            size: 1024,
+        },
+        GetAddress {
+            addr: Location::Global(String::from("testing")),
+            dst: C,
+        },
         CoreOp::put_string(">> ", Output::stdout_char()),
         Push(C, 1),
         CallLabel(String::from("getstr")),
